refactor: Extract shouldReturnResultWithRequestedYearIfThirdPartyDoesNotProvideYear
