+++ src/test/java/no/sikt/nva/pubchannels/handler/search/BaseSearchSerialPublicationByQueryHandlerTest.java
@@ -80,11 +80,37 @@ public abstract class BaseSearchSerialPublicationByQueryHandlerTest extends Sear
         assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
     }
 
+    @Test
+    void shouldReturnResultWithRequestedYearIfThirdPartyDoesNotProvideYear()
+        throws IOException, UnprocessableContentException {
+        var year = randomYear();
+        var issn = randomIssn();
+        var expectedSearchResult = getExpectedPaginatedSearchResultIssnSearchThirdPartyDoesNotProvideYear(year, issn);
+
+        var input = constructRequest(Map.of("year", year, "query", issn), MediaType.ANY_TYPE);
+
+        this.handlerUnderTest.handleRequest(input, output, context);
+
+        var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
+        var pagesSearchResult = objectMapper.readValue(response.getBody(), TYPE_REF);
+
+        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_OK)));
+        assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
+    }
+
     private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearch(
         String year, String printIssn)
         throws UnprocessableContentException {
         var pid = UUID.randomUUID().toString();
         var testChannel = new TestChannel(year, pid, type).withPrintIssn(printIssn);
+        return mockChannelFoundAndReturnExpectedResponse(year, printIssn, testChannel);
+    }
+
+    private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearchThirdPartyDoesNotProvideYear(
+        String year, String printIssn)
+        throws UnprocessableContentException {
+        var pid = UUID.randomUUID().toString();
+        var testChannel = new TestChannel(null, pid, type).withPrintIssn(printIssn);
         return mockChannelFoundAndReturnExpectedResponse(year, printIssn, testChannel);
     }
