+++ src/test/java/no/sikt/nva/pubchannels/handler/search/journal/SearchJournalByQueryHandlerTest.java
@@ -1,14 +1,11 @@
 package no.sikt.nva.pubchannels.handler.search.journal;
 
-import static java.util.Objects.nonNull;
 import static no.sikt.nva.pubchannels.TestConstants.API_DOMAIN;
 import static no.sikt.nva.pubchannels.TestConstants.CHANNEL_REGISTRY_PAGE_COUNT_PARAM;
 import static no.sikt.nva.pubchannels.TestConstants.CHANNEL_REGISTRY_PAGE_NO_PARAM;
 import static no.sikt.nva.pubchannels.TestConstants.CUSTOM_DOMAIN_BASE_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_OFFSET;
-import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_OFFSET_INT;
 import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_SIZE;
-import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_SIZE_INT;
 import static no.sikt.nva.pubchannels.TestConstants.JOURNAL_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.JOURNAL_TYPE;
 import static no.sikt.nva.pubchannels.TestConstants.NAME_QUERY_PARAM;
@@ -20,7 +17,6 @@ import static no.sikt.nva.pubchannels.handler.TestUtils.getChannelRegistrySearch
 import static no.sikt.nva.pubchannels.handler.TestUtils.getChannelRegistrySearchResult;
 import static no.sikt.nva.pubchannels.handler.TestUtils.randomYear;
 import static no.unit.nva.testutils.RandomDataGenerator.objectMapper;
-import static no.unit.nva.testutils.RandomDataGenerator.randomIssn;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static nva.commons.core.attempt.Try.attempt;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -36,10 +32,8 @@ import java.net.HttpURLConnection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import no.sikt.nva.pubchannels.channelregistry.ChannelType;
 import no.sikt.nva.pubchannels.channelregistry.model.ChannelRegistrySerialPublication;
-import no.sikt.nva.pubchannels.handler.TestChannel;
 import no.sikt.nva.pubchannels.handler.ThirdPartySerialPublication;
 import no.sikt.nva.pubchannels.handler.model.SerialPublicationDto;
 import no.sikt.nva.pubchannels.handler.search.BaseSearchSerialPublicationByQueryHandlerTest;
@@ -66,24 +60,6 @@ class SearchJournalByQueryHandlerTest extends BaseSearchSerialPublicationByQuery
                                      .addChild(CUSTOM_DOMAIN_BASE_PATH)
                                      .addChild(JOURNAL_PATH)
                                      .getUri();
-    }
-
-    @Test
-    void shouldReturnResultWithRequestedYearIfThirdPartyDoesNotProvideYear()
-        throws IOException, UnprocessableContentException {
-        var year = randomYear();
-        var issn = randomIssn();
-        var expectedSearchResult = getExpectedPaginatedSearchResultIssnSearchThirdPartyDoesNotProvideYear(year, issn);
-
-        var input = constructRequest(Map.of("year", year, "query", issn), MediaType.ANY_TYPE);
-
-        this.handlerUnderTest.handleRequest(input, output, context);
-
-        var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
-        var pagesSearchResult = objectMapper.readValue(response.getBody(), TYPE_REF);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_OK)));
-        assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
     }
 
     @Test
@@ -261,37 +237,5 @@ class SearchJournalByQueryHandlerTest extends BaseSearchSerialPublicationByQuery
 
     private static SerialPublicationDto toJournalResult(ThirdPartySerialPublication journal, String requestedYear) {
         return SerialPublicationDto.create(constructPublicationChannelUri(JOURNAL_PATH, null), journal, requestedYear);
-    }
-
-    private PaginatedSearchResult<SerialPublicationDto> getExpectedSearchResult(String year,
-                                                                                       String printIssn,
-                                                                                       TestChannel testChannel)
-        throws UnprocessableContentException {
-        var expectedParams = new HashMap<String, String>();
-        expectedParams.put("query", printIssn);
-        if (nonNull(year)) {
-            expectedParams.put("year", year);
-        }
-
-        var expectedHits = List.of(testChannel.asSerialPublicationDto(selfBaseUri, year));
-
-        return PaginatedSearchResult.create(constructPublicationChannelUri(JOURNAL_PATH, expectedParams),
-                                            DEFAULT_OFFSET_INT,
-                                            DEFAULT_SIZE_INT,
-                                            expectedHits.size(),
-                                            expectedHits);
-    }
-
-    private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearchThirdPartyDoesNotProvideYear(
-        String year,
-        String printIssn) throws UnprocessableContentException {
-        var pid = UUID.randomUUID().toString();
-        var testChannel = new TestChannel(null, pid, "Journal").withPrintIssn(printIssn);
-
-        mockChannelRegistryResponse(String.valueOf(year),
-                                    printIssn,
-                                    List.of(testChannel.asChannelRegistrySerialPublicationBody()));
-
-        return getExpectedSearchResult(year, printIssn, testChannel);
     }
 }
